+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Add `obj0` (matches any object) and `empty` (matches the empty object)
   combinators.
 - Add the `union` combinator.
+- Add the `dft` field combinator.
 
 # 2.0.0 (2024-01-12)